core: Move the `Matcher` trait
This is in preparation of moving `Backend`, `Index` and `Store`. I've also added the `Nothing/EverythingMatcher`
even though it's a clear layering violation but I've granted myself an exception for that since they're quite basic.
Its a simple move since it only depends on `RepoPathBuf`
which already is in the new core crate.

Part of #6284
