+++ lib/src/matchers.rs
@@ -19,115 +19,16 @@ use std::fmt::Debug;
 
 use globset::Glob;
 use itertools::Itertools as _;
+pub use jj_core::matchers::EverythingMatcher;
+pub use jj_core::matchers::Matcher;
+pub use jj_core::matchers::NothingMatcher;
+pub use jj_core::matchers::Visit;
+pub use jj_core::matchers::VisitDirs;
+pub use jj_core::matchers::VisitFiles;
 use tracing::instrument;
 
 use crate::repo_path::RepoPath;
-use crate::repo_path::RepoPathComponentBuf;
 use crate::repo_path::RepoPathTree;
-
-#[derive(PartialEq, Eq, Debug)]
-pub enum Visit {
-    /// Everything in the directory is *guaranteed* to match, no need to check
-    /// descendants
-    AllRecursively,
-    Specific {
-        dirs: VisitDirs,
-        files: VisitFiles,
-    },
-    /// Nothing in the directory or its subdirectories will match.
-    ///
-    /// This is the same as `Specific` with no directories or files. Use
-    /// `Visit::set()` to get create an instance that's `Specific` or
-    /// `Nothing` depending on the values at runtime.
-    Nothing,
-}
-
-impl Visit {
-    /// All entries in the directory need to be visited, but they are not
-    /// guaranteed to match.
-    const SOME: Self = Self::Specific {
-        dirs: VisitDirs::All,
-        files: VisitFiles::All,
-    };
-
-    fn sets(dirs: HashSet<RepoPathComponentBuf>, files: HashSet<RepoPathComponentBuf>) -> Self {
-        if dirs.is_empty() && files.is_empty() {
-            Self::Nothing
-        } else {
-            Self::Specific {
-                dirs: VisitDirs::Set(dirs),
-                files: VisitFiles::Set(files),
-            }
-        }
-    }
-
-    pub fn is_nothing(&self) -> bool {
-        *self == Self::Nothing
-    }
-}
-
-#[derive(PartialEq, Eq, Debug)]
-pub enum VisitDirs {
-    All,
-    Set(HashSet<RepoPathComponentBuf>),
-}
-
-#[derive(PartialEq, Eq, Debug)]
-pub enum VisitFiles {
-    All,
-    Set(HashSet<RepoPathComponentBuf>),
-}
-
-pub trait Matcher: Debug + Send + Sync {
-    fn matches(&self, file: &RepoPath) -> bool;
-    fn visit(&self, dir: &RepoPath) -> Visit;
-}
-
-impl<T: Matcher + ?Sized> Matcher for &T {
-    fn matches(&self, file: &RepoPath) -> bool {
-        <T as Matcher>::matches(self, file)
-    }
-
-    fn visit(&self, dir: &RepoPath) -> Visit {
-        <T as Matcher>::visit(self, dir)
-    }
-}
-
-impl<T: Matcher + ?Sized> Matcher for Box<T> {
-    fn matches(&self, file: &RepoPath) -> bool {
-        <T as Matcher>::matches(self, file)
-    }
-
-    fn visit(&self, dir: &RepoPath) -> Visit {
-        <T as Matcher>::visit(self, dir)
-    }
-}
-
-#[derive(PartialEq, Eq, Debug)]
-pub struct NothingMatcher;
-
-impl Matcher for NothingMatcher {
-    fn matches(&self, _file: &RepoPath) -> bool {
-        false
-    }
-
-    fn visit(&self, _dir: &RepoPath) -> Visit {
-        Visit::Nothing
-    }
-}
-
-#[derive(PartialEq, Eq, Debug)]
-pub struct EverythingMatcher;
-
-impl Matcher for EverythingMatcher {
-    fn matches(&self, _file: &RepoPath) -> bool {
-        true
-    }
-
-    fn visit(&self, _dir: &RepoPath) -> Visit {
-        Visit::AllRecursively
-    }
-}
 
 #[derive(PartialEq, Eq, Debug)]
 pub struct FilesMatcher {
@@ -527,6 +428,7 @@ mod tests {
 
     use super::*;
     use crate::fileset::parse_file_glob;
+    use crate::repo_path::RepoPathComponentBuf;
 
     fn repo_path(value: &str) -> &RepoPath {
         RepoPath::from_internal_string(value).unwrap()
